requirements-colab.txt
ultralytics>=8.3.0
open_clip_torch>=2.24.0
faiss-cpu>=1.8.0
opencv-python-headless>=4.10.0
scikit-learn>=1.4.0
tqdm>=4.66.0
kaggle>=1.6.17
Pillow>=10.0.0
